feat: 直线添加参数式方程内部类，面添加一般式方程内部类，新增计算类，用于三维空间各种计算，视点类新加两个向量，所有所有表示坐标的数值改为double

+++ src/main/java/com/hysea/entity/Vector.java
@@ -2,39 +2,39 @@ package com.hysea.entity;
 
 public class Vector {
 
-    private int x;
+    private double x;
 
-    private int y;
+    private double y;
 
-    private int z;
+    private double z;
 
-    public Vector(int x, int y, int z) {
+    public Vector(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
 }
